feature: write unit-test for project component
