+++ src/components/Projects/Carousel/SwiperSlider.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import ProjectCard from "../ProjectCard/ProjectCard";
-import style from "../projects.module.css";
+import style from "../Projects.module.css";
 import "swiper/css";
 import "swiper/css/navigation";
 import "swiper/css/effect-coverflow";
