+++ src/components/UI/Button/Button.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import style from "./button.module.css";
+import style from "./Button.module.css";
 
 const Button = ({
   variant = "large",
